Tests, documentation and more
Introduces tests for the features and fixed some stuff for all tests to be passing!
Yup all tests are passing.

Added documentation in Doxygen, and added documentation to every class. Created some seperate custom pages for introduction. The docs should be put into the github pages, I'll work on that after this.

Besides all this, I tinkered a bit with the design and changes some stuff.

Well there are a lot of changes, definitely good for one commit 💀 Well half of it is testing framework and docs.