Hotfix positioning and tests
Added a fix to the typo which caused the parent node to update global position instead of the child node.

Fixed tests so this doesn't happen unnoticed again.
In the tests somehow I didn't notice the position wasn't being set, probably due to copying.

All tests are passing now, and positioning is well.