Hover/Focus styles, and Merging.
Adds the Hover and Focus styles and style merging.

Hover and focus styles are dependent on the normal style, whenever it contains an unset value, the normal style fills it in.
Merging of the styles does exactly that, it fills the unset values from the primary style.

The primary style for focus when hovering is the hover style merged with normal style.

Other changes:
Changed some of the names to better reflect the properties.
Updated the documentation to reflect the changes.
Added new documentation.
Ignored and removed GDUnit test reports which regenerate each time and clutter real changes.
Created special resources for existing visual tests, so that changes in the library doesn't reset the styles.